Improvement

+++ app/src/main/java/com/f3ffo/hellobusbologna/hellobus/DownloadCsv.java
@@ -7,20 +7,12 @@ import android.util.Log;
 
 import org.apache.commons.io.FileUtils;
 
-import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
-import java.io.InputStreamReader;
-import java.net.HttpURLConnection;
 import java.net.URL;
 import java.nio.channels.Channels;
 import java.nio.channels.ReadableByteChannel;
-import java.nio.charset.StandardCharsets;
-import java.util.Objects;
-
-import okhttp3.OkHttpClient;
-import okhttp3.Request;
 
 public class DownloadCsv extends AsyncTask<Void, Void, Void> {
 
@@ -28,63 +20,40 @@ public class DownloadCsv extends AsyncTask<Void, Void, Void> {
     private Context context;
     private String version;
 
-    public DownloadCsv(Context context) {
+    public DownloadCsv(Context context, String version) {
         this.context = context;
-    }
-
-    private void checkFile() {
-        try {
-            Request get = new Request.Builder().url("https://solweb.tper.it/web/tools/open-data/open-data-download.aspx?source=solweb.tper.it&filename=opendata-versione&version=1&format=csv").build();
-            BufferedReader br = new BufferedReader(new InputStreamReader(Objects.requireNonNull(new OkHttpClient().newCall(get).execute().body()).byteStream(), StandardCharsets.UTF_8));
-            String versionUpdate;
-            do {
-                versionUpdate = br.readLine();
-            } while (!versionUpdate.startsWith("lineefermate"));
-            this.version = versionUpdate.substring(versionUpdate.lastIndexOf(";") + 1);
-        } catch (IOException e) {
-            Log.e("ERROR checkFile01", e.getMessage());
-        }
-        try {
-            FileUtils.touch(new File(context.getFilesDir(), "cut_" + version + ".csv"));
-            FileUtils.touch(new File(context.getFilesDir(), "favourites.properties"));
-        } catch (IOException e) {
-            Log.e("ERROR checkFile02", e.getMessage());
-        }
+        this.version = version;
     }
 
     @Override
     protected Void doInBackground(Void... voids) {
-        checkFile();
         File[] listFiles = context.getFilesDir().listFiles();
-        for (File listFile : listFiles) {
-            if (!listFile.isDirectory() && !listFile.getName().contains(this.version) && !listFile.getName().equals("favourites.properties")) {
-                try {
-                    FileUtils.forceDelete(new File(context.getFilesDir(), listFile.getName()));
-                    Log.i("FILE_DELETED", listFile.getName());
-                } catch (IOException e) {
-                    Log.e("ERROR FILE_DELETE", listFile.getName());
+        if (listFiles != null) {
+            for (File listFile : listFiles) {
+                if (!listFile.isDirectory() && !listFile.getName().contains(this.version) && !listFile.getName().equals("favourites.properties")) {
+                    try {
+                        FileUtils.forceDelete(new File(context.getFilesDir(), listFile.getName()));
+                        Log.i("FILE_DELETED", listFile.getName());
+                    } catch (IOException e) {
+                        Log.e("ERROR FILE_DELETE", listFile.getName());
+                    }
                 }
             }
         }
         if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.O) {
-            ReadableByteChannel readableByteChannel = null;
-            FileOutputStream outputStream = null;
+            ReadableByteChannel readableByteChannel;
+            FileOutputStream outputStream;
             try {
                 URL url = new URL("https://solweb.tper.it/web/tools/open-data/open-data-download.aspx?source=solweb.tper.it&filename=lineefermate&version=" + version + "&format=csv");
                 readableByteChannel = Channels.newChannel(url.openStream());
                 outputStream = context.openFileOutput("lineefermate_" + version + ".csv", Context.MODE_PRIVATE);
                 outputStream.getChannel().transferFrom(readableByteChannel, 0, Long.MAX_VALUE);
+                outputStream.close();
+                readableByteChannel.close();
+                Log.i("FILE SCARICATO", "PORCA MADONNA");
             } catch (IOException e) {
                 Log.e("ERROR fileDownload_O", e.getMessage());
-            } finally {
-                try {
-                    if (outputStream != null && readableByteChannel != null) {
-                        outputStream.close();
-                        readableByteChannel.close();
-                    }
-                } catch (IOException e) {
-                    Log.e("ERROR closeDownload", e.getMessage());
-                }
+                e.printStackTrace();
             }
         } else {
             try {
@@ -92,6 +61,7 @@ public class DownloadCsv extends AsyncTask<Void, Void, Void> {
                 FileUtils.copyURLToFile(url, new File(context.getFilesDir() + "/lineefermate_" + version + ".csv"));
             } catch (IOException e) {
                 Log.e("ERROR fileDownload_M", e.getMessage());
+                e.printStackTrace();
             }
         }
         return null;
